+++ docs/plugins/meta.ts
@@ -1,0 +1,8 @@
+import { defineMeta } from 'blume'
+
+export default defineMeta({
+	title: 'Integrationen',
+	order: 40,
+	icon: 'heart-handshake',
+	collapsed: false,
+})