+++ docs/user/meta.ts
@@ -1,0 +1,8 @@
+import { defineMeta } from 'blume'
+
+export default defineMeta({
+	title: 'Benutzer-Guides',
+	order: 10,
+	icon: 'rocket',
+	collapsed: false,
+})